Resolve circular dependency problem
